fix(mdx): correctly recognize inline JSX (#5783)
